fix(gatsby-ssr): can't export a function unless it's in the API, so to utilities checkPathExclusion
