+++ src/gatsby-ssr.spec.ts
@@ -1,4 +1,4 @@
-import { onPreRenderHTML, onRenderBody, checkPathExclusion } from './gatsby-ssr'
+import { onPreRenderHTML, onRenderBody } from './gatsby-ssr'
 import Spy = jasmine.Spy
 import { headComponentsData, postBodyComponentsData, scriptsData } from './fake-data.spec'
 
@@ -206,16 +206,6 @@ describe('gatsby-ssr.js', () => {
 
       expect(replaceHeadComponentsSpy).toHaveBeenCalledTimes(0)
       expect(replacePostBodyComponentsSpy).toHaveBeenCalledTimes(0)
-    })
-  })
-
-  describe('checkPathExclusion', () => {
-    it('does the pathname match any of the exclusions', () => {
-      expect(checkPathExclusion('/client/omegablerns', { excludePaths: /\/client/ })).toBeTruthy()
-      expect(checkPathExclusion('/my-cool-page', { excludePaths: /\/my-cool-page/ })).toBeTruthy()
-      expect(checkPathExclusion('/about/blerns', { excludePaths: /(\/client)|(\/tacos)/ })).toBeFalsy()
-      expect(checkPathExclusion('/about/tacos', { excludePaths: /(\/client)|(\/about\/tacos)/ })).toBeTruthy()
-      expect(checkPathExclusion('/blerkstorm', {})).toBeFalsy()
     })
   })
 })
